+++ package.json
@@ -7,8 +7,14 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
+    "@types/node": "^20.8.6",
+    "@types/react": "^18.2.28",
+    "@types/react-dom": "^18.2.13",
+    "@typescript-eslint/eslint-plugin": "^6.8.0",
+    "@typescript-eslint/parser": "^6.8.0",
     "axios": "^1.5.1",
     "bootstrap": "^5.3.2",
+    "css-loader": "^6.8.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.9.0",
     "react-daum-postcode": "^3.1.3",
@@ -21,8 +27,10 @@
     "react-select": "^5.7.7",
     "recoil": "^0.7.7",
     "sass": "^1.68.0",
+    "style-loader": "^3.3.3",
     "styled-components": "^6.0.8",
     "styled-reset": "^4.5.1",
+    "typescript": "^5.2.2",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
