+++ src/global.d.ts
@@ -1,0 +1,7 @@
+declare global {
+  interface Window {
+    kakao: any
+  }
+}
+
+export {}
