--- src/index.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import './index.css'
-import App from './App'
-import reportWebVitals from './reportWebVitals'
-import 'bootstrap/dist/css/bootstrap.css'
-
-const root = ReactDOM.createRoot(document.getElementById('root'))
-root.render(<App />)
-
-reportWebVitals()
