--- src/pages/NotFound.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const NotFound = () => {
-  return <div>NotFound</div>;
-};
-
-export default NotFound;
