+++ src/pages/NotFound.tsx
@@ -1,0 +1,5 @@
+const NotFound = () => {
+  return <div>NotFound</div>
+}
+
+export default NotFound
