+++ src/router/Router.tsx
@@ -2,9 +2,9 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import GlobalStyle from '../layout/GlobalStyle'
 import Main from '../pages/main/MainPage'
 import RegisterPage from '../pages/register/RegisterPage'
-import Apart from './../pages/apart/Apart'
-import Studio from './../pages/studio/Studio'
-import Office from './../pages/office/Office'
+import Apart from '../pages/apart/Apart'
+import Studio from '../pages/studio/Studio'
+import Office from '../pages/office/Office'
 import MyPage from '../pages/my/MyPage'
 import NotFound from '../pages/NotFound'
