+++ src/util/generateMap.ts
@@ -1,4 +1,6 @@
-const generateMap = (element, options = null) => {
+import { ReactNode } from 'react'
+
+const generateMap = (element: ReactNode, options = null) => {
   const defaultOptions = {
     center: new window.kakao.maps.LatLng(37.498086, 127.028001),
     level: 3,
