+++ tsconfig.json
@@ -1,0 +1,31 @@
+{
+  "compilerOptions": {
+    "experimentalDecorators": true,
+    "allowJs": true, // TypeScript 컴파일러가 JavaScript 파일도 컴파일
+    "alwaysStrict": true,
+    "esModuleInterop": true, // CommonJS 모듈을 ES6 모듈로 변환할 때 import 구문을 사용
+    "downlevelIteration": true, // ES6 이터레이터를 ES5 코드로 변환할 때 이터레이터 코드를 낮은 버전의 JavaScript 엔진에서도 실행
+    "forceConsistentCasingInFileNames": true,
+    "isolatedModules": true, // 각 파일이 독립된 모듈로 처리
+    "jsx": "react-jsx", // JSX 구문을 사용
+    "lib": ["DOM.Iterable", "DOM", "ES2020", "ESNext"], // 사용할 라이브러리
+    "module": "ESNext", // 사용할 모듈 시스템
+    "moduleResolution": "Bundler", // 모듈 해석 방식
+    "noEmit": true, // 컴파일 결과를 출력하지 않도록 지정
+    "noFallthroughCasesInSwitch": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "allowSyntheticDefaultImports": true,
+    "sourceMap": true,
+    // "paths": {
+    //   "@/*": ["src/*"]
+    // },
+    "resolveJsonModule": true,
+    "skipLibCheck": false, // 선언 파일 (.d.ts 파일)의 타입 체크를 비활성화. 컴파일 시간이 더 빨라짐
+    "strict": true,
+    "target": "ES2020",
+    "incremental": true // 증분 컴파일을 활성화하여 컴파일 속도를 높임
+  },
+  "exclude": ["node_modules"], // 컴파일하지 않을 파일 또는 디렉토리를 지정
+  "include": ["./src"] // 컴파일 할 파일 또는 디렉토리를 지정
+}
